+++ packages/peregrine/lib/talons/CheckoutPage/PaymentInformation/clearPaymentDetailsCollection.js
@@ -1,0 +1,9 @@
+import useClearBrainTreeDetails from './useClearBrainTreeDetails';
+
+/**
+ * TODO must be a target to allow payments to define own clear methods what get call when
+ * placeOrder throw an error
+ */
+export const clearPaymentDetailsCollection = {
+    braintree: useClearBrainTreeDetails
+};
