+++ packages/peregrine/lib/talons/CheckoutPage/PaymentInformation/useClearBrainTreeDetails.js
@@ -1,0 +1,25 @@
+const useClearBrainTreeDetails = () => {
+
+     const operations = mergeOperations(
+         DEFAULT_OPERATIONS,
+         props.operations
+     );
+    const [{ cartId }] = useCartContext();
+    const client = useApolloClient();
+    const clearPaymentDetails = useCallback(() => {
+        client.writeQuery({
+            query: getPaymentNonceQuery,
+            data: {
+                cart: {
+                    __typename: 'Cart',
+                    id: cartId,
+                    paymentNonce: null
+                }
+            }
+        });
+    }, [cartId, client, getPaymentNonceQuery]);
+
+    return { clearPaymentDetails };
+};
+
+export default useClearBrainTreeDetails;
