+++ packages/peregrine/lib/targets/__tests__/peregrine-targets.spec.js
@@ -133,8 +133,7 @@ test('exposes all hooks and targets', async () => {
           talons.CheckoutPage.ItemsReview.useItemsReview.wrapWith() wraps export "useItemsReview" from "CheckoutPage/ItemsReview/useItemsReview.js"
           talons.CheckoutPage.OrderConfirmationPage.useCreateAccount.wrapWith() wraps export "useCreateAccount" from "CheckoutPage/OrderConfirmationPage/useCreateAccount.js"
           talons.CheckoutPage.OrderConfirmationPage.useOrderConfirmationPage.wrapWith() wraps export "useOrderConfirmationPage" from "CheckoutPage/OrderConfirmationPage/useOrderConfirmationPage.js"
-          talons.CheckoutPage.PaymentInformation.useBraintreeSummary.wrapWith() wraps export "useBraintreeSummary" from "CheckoutPage/PaymentInformation/useBraintreeSummary.js"
-          talons.CheckoutPage.PaymentInformation.useCreditCard.wrapWith() wraps export "useCreditCard" from "CheckoutPage/PaymentInformation/useCreditCard.js"
+          talons.CheckoutPage.PaymentInformation.useClearBrainTreeDetails.wrapWith() wraps export "useClearBrainTreeDetails" from "CheckoutPage/PaymentInformation/useClearBrainTreeDetails.js"
           talons.CheckoutPage.PaymentInformation.useEditModal.wrapWith() wraps export "useEditModal" from "CheckoutPage/PaymentInformation/useEditModal.js"
           talons.CheckoutPage.PaymentInformation.usePaymentInformation.wrapWith() wraps export "usePaymentInformation" from "CheckoutPage/PaymentInformation/usePaymentInformation.js"
           talons.CheckoutPage.PaymentInformation.usePaymentMethods.wrapWith() wraps export "usePaymentMethods" from "CheckoutPage/PaymentInformation/usePaymentMethods.js"
